+++ app/config/deploy/prod.rb
@@ -6,7 +6,7 @@ set :user,      "www-data"
 set :deploy_via, :copy
 
 # Options to pass to composer when installing/updating
-set :composer_options, " --no-interaction --verbose --prefer-dist --optimize-autoloader --no-progress --ignore-platform-reqs"
+set :composer_options, " --no-interaction --verbose --prefer-dist --optimize-autoloader --no-progress"
 
 # Clear *_dev controllers
 set :clear_controllers, true